modules/ai-agents/requirements.txt
# AI Agents Module Dependencies
# Framework
agno>=0.1.0
mcp-agent>=0.1.0

# Vector Database
lancedb>=0.4.0

# LLM APIs
openai>=1.0.0
google-generativeai>=0.3.0
google-genai>=1.0.0

# Web Search
serpapi>=0.1.0

# Web Framework (for API endpoints)
fastapi>=0.109.0
uvicorn>=0.27.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.26.0

# Streamlit (for admin interfaces)
streamlit>=1.30.0

# Browser Automation (for MCP)
playwright>=1.41.0

# Scheduling
schedule>=1.2.0
